version 1.3 released
git-svn-id: https://plugins.svn.wordpress.org/documentor-lite/trunk@1293516 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ uninstall.php
@@ -1,4 +1,8 @@
 <?php 
+// If uninstall is not called from WordPress, exit
+if ( !defined( 'WP_UNINSTALL_PLUGIN' ) ) {
+    exit();
+}
 delete_option('documentorlite_db_version');
 if( !defined( 'DOCUMENTOR_PLUGIN_BASENAME' ) ) {
 	global $wpdb, $table_prefix;
